requirements.txt
twisted==13.2.0
Flask==0.10.1
Flask-SQLAlchemy==1.0
requests==2.2.1
python-dateutil==2.2
coverage==3.7.1